Use the iterate method instead of the all method to get `table_data`
[`iterate`](https://pyairtable.readthedocs.io/en/latest/api.html?highlight=all#pyairtable.api.Table.iterate) fetches records in pages,
where each page fetched is a list, and `iterate`'s return value is a generator of those list/pages.

`all` concatenates each of those pages of records into a single eagerly-evaluated list.

This commit replaces the use of `all` with a call to `iterate` in a generator comprehension to flatten the generator.